あと一息，なぜかImportノードがロードできない．
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@2392 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Framework/FWSdk.h
@@ -56,7 +56,7 @@ public:
 	virtual void SetDebugMode(bool debug = true){debugMode = debug;}
 
 	virtual size_t NChildObject() const { return NScene(); }
-	virtual ObjectIf* GetChildObject(size_t i){ return GetScene(i); }
+	virtual ObjectIf* GetChildObject(size_t i){ return GetScene((int)i); }
 	virtual bool AddChildObject(ObjectIf* o);
 	virtual PHSdkIf* GetPHSdk(){ return phSdk; }
 	virtual GRSdkIf* GetGRSdk(){ return grSdk; }
